requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

paho-mqtt==1.6.1
git+https://gitlab.mi.hdm-stuttgart.de/iotee/python-iotee.git
